+++ inte_subject/forms/family_history_form.py
@@ -1,12 +1,9 @@
 from django import forms
 from edc_form_validators.form_validator import FormValidator
+from respond_forms.utils import raise_if_clinical_review_does_not_exist
 
 from ..models import FamilyHistory
-from .mixins import (
-    CrfFormValidatorMixin,
-    CrfModelFormMixin,
-    raise_if_clinical_review_does_not_exist,
-)
+from .mixins import CrfFormValidatorMixin, CrfModelFormMixin
 
 
 class FamilyHistoryFormValidator(CrfFormValidatorMixin, FormValidator):