+++ inte_subject/forms/integrated_care_review_form.py
@@ -1,12 +1,12 @@
 from django import forms
 from edc_constants.constants import OTHER, YES
 from edc_form_validators import FormValidator
+from respond_forms.utils import raise_if_clinical_review_does_not_exist
 
 from ..models import IntegratedCareReview
 from .mixins import (
     CrfFormValidatorMixin,
     CrfModelFormMixin,
-    raise_if_clinical_review_does_not_exist,
     raise_if_intervention_site_without_icc_registration,
 )